added validation with mongoose for createPost

+++ server/src/resources/posts/post-controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 import "express-async-errors";
-import { Types as MongooseTypes } from "mongoose";
+import { Error, Types as MongooseTypes } from "mongoose";
 import { PostModel } from "./post-model";
 
 export async function getAllPosts(req: Request, res: Response) {
@@ -23,16 +23,23 @@ export async function getPostById(req: Request, res: Response) {
 }
 
 export async function createPost(req: Request, res: Response) {
-  // validate body
-  // ensure user is logged in
   if (!req.session?.user) {
     res.status(401).json("You must login to create a post in your username");
     return;
   }
-  const postData = { ...req.body, author: req.session.user };
-  const post = new PostModel(postData);
-  await post.save();
-  res.status(201).json(post);
+
+  try {
+    const postData = { ...req.body, author: req.session.user };
+    const post = new PostModel(postData);
+    await post.save();
+    res.status(201).json(post);
+  } catch (error) {
+    if ((error as Error.ValidationError).name === "ValidationError") {
+      res.status(400).json((error as Error.ValidationError).message);
+    } else {
+      res.status(500).json("An unexpected error occurred.");
+    }
+  }
 }
 
 export async function updatePost(req: Request, res: Response) {
@@ -44,7 +51,7 @@ export async function updatePost(req: Request, res: Response) {
     author: userId,
   });
   if (!existingPost) {
-    return res.status(404).json("Post not found or not owned by the user");
+    return res.status(403).json("Post not found or not owned by the user");
   }
   // Update the post
   const updatedPost = await PostModel.findByIdAndUpdate(postId, req.body, {
@@ -56,14 +63,21 @@ export async function updatePost(req: Request, res: Response) {
 export async function deletePost(req: Request, res: Response) {
   const postId = req.params.id;
   const userId = req.session!.userId;
-  // Check if the post exists and belongs to the user
-  const existingPost = await PostModel.findOne({
-    _id: postId,
-    author: userId,
-  });
+
+  // Check if the post exists
+  const existingPost = await PostModel.findById(postId);
+
   if (!existingPost) {
-    return res.status(404).json("Post not found or not owned by the user");
+    return res.status(404).json("Post not found");
   }
+
+  // Check if the post belongs to the user
+  if (existingPost.author.toString() !== userId) {
+    return res
+      .status(403)
+      .json("You do not have permission to delete this post");
+  }
+
   // Delete the post
   await PostModel.findByIdAndDelete(postId);
   res.status(204).json({ message: "Post deleted successfully" });
